Created azure_storage_datalake crate from azure_storage
