+++ .github/workflows/build.yml
@@ -143,7 +143,7 @@ jobs:
 
     - name: e2e tests build
       run: |
-        PROJECTS=(core  cosmos  identity  messaging_servicebus  storage  storage_blobs  storage_queues  data_tables)
+        PROJECTS=(core  cosmos  identity  messaging_servicebus  storage  storage_blobs  storage_queues  storage_datalake  data_tables)
         for PROJ in ${PROJECTS[@]}
         do
           echo "Checking e2e tests for $PROJ"
