--- sdk/storage/src/data_lake/prelude.rs
@@ -1,3 +1,0 @@
-pub use crate::data_lake::clients::*;
-pub use crate::data_lake::operations::*;
-pub use crate::data_lake::Properties;
