+++ sdk/storage_datalake/Cargo.toml
@@ -1,0 +1,34 @@
+[package]
+name = "azure_storage_datalake"
+version = "0.1.0"
+description = "Rust wrappers around Microsoft Azure Data Lake Storage REST APIs"
+readme = "README.md"
+authors = ["Microsoft Corp."]
+license = "MIT"
+repository = "https://github.com/azure/azure-sdk-for-rust"
+homepage = "https://github.com/azure/azure-sdk-for-rust"
+documentation = "https://docs.rs/azure_storage_datalake"
+keywords = ["sdk", "azure", "rest", "iot", "cloud", "storage", "datalake"]
+categories = ["api-bindings"]
+edition = "2018"
+
+[dependencies]
+async-trait = "0.1"
+azure_core = { path = "../core", version = "0.1.0" }
+azure_storage = { path = "../storage", version = "0.1.0" }
+base64 = "0.13"
+bytes = "1.0"
+chrono = { version = "0.4", features = ["serde"] }
+futures = "0.3"
+http = "0.2"
+log = "0.4"
+serde = { version = "1.0" }
+serde_derive = "1.0"
+serde_json = "1.0"
+serde-xml-rs = "0.4"
+uuid = { version = "0.8", features = ["v4"] }
+url = "2.2"
+
+[dev-dependencies]
+tokio = { version = "1.0", features = ["full"] }
+azure_identity = { path = "../identity", version = "0.1.0" }
