+++ sdk/storage_datalake/README.md
@@ -1,0 +1,12 @@
+# Azure SDK for Rust - Azure Data Lake Storage
+
+Azure Data Lake Storage crate for the unofficial Microsoft Azure SDK for Rust. This crate is part of a collection of crates: for more information please refer to [https://github.com/azure/azure-sdk-for-rust](https://github.com/azure/azure-sdk-for-rust).
+
+## Usage
+
+To set this crate as a dependency, add this to your Cargo.toml
+
+```toml
+[dependencies]
+azure_storage_datalake = { version = "0.1.0", git = "https://github.com/Azure/azure-sdk-for-rust" }
+```
