+++ sdk/storage_datalake/examples/data_lake_03_file_rename.rs
@@ -2,7 +2,7 @@ use azure_core::prelude::*;
 use azure_identity::token_credentials::DefaultAzureCredential;
 use azure_identity::token_credentials::TokenCredential;
 use azure_storage::core::prelude::*;
-use azure_storage::data_lake::prelude::*;
+use azure_storage_datalake::prelude::*;
 use chrono::Utc;
 use std::error::Error;
 
@@ -19,34 +19,43 @@ async fn main() -> Result<(), Box<dyn Error + Send + Sync>> {
     let create_fs_response = file_system_client.create().execute().await?;
     println!("create file system response == {:?}\n", create_fs_response);
 
-    let file_path = "some/path/example-file.txt";
+    let file_path1 = "some/path/example-file1.txt";
+    let file_path2 = "some/path/example-file2.txt";
 
-    println!("creating file '{}'...", file_path);
-    let create_file_response = file_system_client
-        .create_file(Context::default(), file_path, FileCreateOptions::default())
+    println!("creating file '{}'...", file_path1);
+    let create_file_response1 = file_system_client
+        .create_file(Context::default(), file_path1, FileCreateOptions::default())
         .await?;
-    println!("create file response == {:?}\n", create_file_response);
+    println!("create file response == {:?}\n", create_file_response1);
 
-    println!("creating file '{}' if not exists...", file_path);
-    let create_file_if_not_exists_result = file_system_client
-        .create_file_if_not_exists(Context::default(), file_path)
+    println!("creating file '{}'...", file_path2);
+    let create_file_response2 = file_system_client
+        .create_file(Context::default(), file_path2, FileCreateOptions::default())
+        .await?;
+    println!("create file response == {:?}\n", create_file_response2);
+
+    println!(
+        "renaming file '{}' to '{}' if not exists...",
+        file_path1, file_path2
+    );
+    let rename_file_if_not_exists_result = file_system_client
+        .rename_file_if_not_exists(Context::default(), file_path1, file_path2)
         .await;
     println!(
-        "create file result (should fail) == {:?}\n",
-        create_file_if_not_exists_result
+        "rename file result (should fail) == {:?}\n",
+        rename_file_if_not_exists_result
     );
 
-    println!("creating file '{}' (overwrite)...", file_path);
-    let create_file_response = file_system_client
-        .create_file(Context::default(), file_path, FileCreateOptions::default())
+    println!("renaming file '{}' to '{}'...", file_path1, file_path2);
+    let rename_file_response = file_system_client
+        .rename_file(
+            Context::default(),
+            file_path1,
+            file_path2,
+            FileRenameOptions::default(),
+        )
         .await?;
-    println!("create file response == {:?}\n", create_file_response);
-
-    println!("deleting file '{}'...", file_path);
-    let delete_file_response = file_system_client
-        .delete_file(Context::default(), file_path, FileDeleteOptions::default())
-        .await?;
-    println!("delete_file file response == {:?}\n", delete_file_response);
+    println!("rename file response == {:?}\n", rename_file_response);
 
     println!("deleting file system...");
     let delete_fs_response = file_system_client.delete().execute().await?;
