+++ sdk/storage_datalake/src/clients/file_system_client.rs
@@ -1,6 +1,6 @@
-use crate::data_lake::operations::*;
-use crate::data_lake::requests::*;
-use crate::{data_lake::clients::DataLakeClient, Properties};
+use crate::operations::*;
+use crate::requests::*;
+use crate::{clients::DataLakeClient, Properties};
 use azure_core::prelude::IfMatchCondition;
 use azure_core::Pipeline;
 use azure_core::{Context, HttpClient};
