+++ sdk/storage_datalake/src/lib.rs
@@ -1,3 +1,12 @@
+#[macro_use]
+extern crate log;
+#[macro_use]
+extern crate serde_derive;
+#[macro_use]
+extern crate azure_core;
+
+pub use azure_storage::{Error, Result};
+
 mod authorization_policy;
 pub mod clients;
 mod file_system;
