+++ sdk/storage_datalake/src/operations/file_delete.rs
@@ -1,16 +1,16 @@
-use crate::core::headers::CommonStorageResponseHeaders;
 use azure_core::prelude::ContentLength;
 use azure_core::prelude::IfMatchCondition;
+use azure_storage::core::headers::CommonStorageResponseHeaders;
 use std::convert::TryInto;
 
 use azure_core::{Request as HttpRequest, Response as HttpResponse};
 
 #[derive(Debug, Clone, Default)]
-pub struct FileFlushOptions<'a> {
+pub struct FileDeleteOptions<'a> {
     if_match_condition: Option<IfMatchCondition<'a>>,
 }
 
-impl<'a> FileFlushOptions<'a> {
+impl<'a> FileDeleteOptions<'a> {
     pub fn new() -> Self {
         Self {
             if_match_condition: None,
@@ -23,18 +23,18 @@ impl<'a> FileFlushOptions<'a> {
 
     pub(crate) fn decorate_request(&self, req: &mut HttpRequest) -> Result<(), crate::Error> {
         azure_core::headers::add_optional_header2(&self.if_match_condition, req)?;
-        azure_core::headers::add_mandatory_header2(&ContentLength::new(0), req)?;
+        azure_core::headers::add_mandatory_header2(&ContentLength::new(0), req)?; // Length is required for creating files
 
         Ok(())
     }
 }
 
 #[derive(Debug, Clone)]
-pub struct FileFlushResponse {
+pub struct FileDeleteResponse {
     pub common_storage_response_headers: CommonStorageResponseHeaders,
 }
 
-impl FileFlushResponse {
+impl FileDeleteResponse {
     pub async fn try_from(response: HttpResponse) -> Result<Self, crate::Error> {
         let (_status_code, headers, _pinned_stream) = response.deconstruct();
