+++ sdk/storage_datalake/src/operations/file_flush.rs
@@ -1,17 +1,16 @@
-use crate::core::headers::CommonStorageResponseHeaders;
+use azure_core::prelude::ContentLength;
 use azure_core::prelude::IfMatchCondition;
-use azure_core::prelude::{ContentLength, ContentType};
-use bytes::Bytes;
+use azure_storage::core::headers::CommonStorageResponseHeaders;
 use std::convert::TryInto;
 
 use azure_core::{Request as HttpRequest, Response as HttpResponse};
 
 #[derive(Debug, Clone, Default)]
-pub struct FileAppendOptions<'a> {
+pub struct FileFlushOptions<'a> {
     if_match_condition: Option<IfMatchCondition<'a>>,
 }
 
-impl<'a> FileAppendOptions<'a> {
+impl<'a> FileFlushOptions<'a> {
     pub fn new() -> Self {
         Self {
             if_match_condition: None,
@@ -22,29 +21,20 @@ impl<'a> FileAppendOptions<'a> {
         if_match_condition: IfMatchCondition<'a> => Some(if_match_condition),
     }
 
-    pub(crate) fn decorate_request(
-        &self,
-        req: &mut HttpRequest,
-        bytes: Bytes,
-    ) -> Result<(), crate::Error> {
+    pub(crate) fn decorate_request(&self, req: &mut HttpRequest) -> Result<(), crate::Error> {
         azure_core::headers::add_optional_header2(&self.if_match_condition, req)?;
-        azure_core::headers::add_mandatory_header2(
-            &ContentType::new("application/octet-stream"),
-            req,
-        )?;
-        azure_core::headers::add_mandatory_header2(&ContentLength::new(bytes.len() as i32), req)?;
-        req.set_body(bytes.into());
+        azure_core::headers::add_mandatory_header2(&ContentLength::new(0), req)?;
 
         Ok(())
     }
 }
 
 #[derive(Debug, Clone)]
-pub struct FileAppendResponse {
+pub struct FileFlushResponse {
     pub common_storage_response_headers: CommonStorageResponseHeaders,
 }
 
-impl FileAppendResponse {
+impl FileFlushResponse {
     pub async fn try_from(response: HttpResponse) -> Result<Self, crate::Error> {
         let (_status_code, headers, _pinned_stream) = response.deconstruct();
