+++ sdk/storage_datalake/src/requests/create_file_system_builder.rs
@@ -1,5 +1,5 @@
-use crate::data_lake::responses::*;
-use crate::{data_lake::clients::FileSystemClient, Properties};
+use crate::responses::*;
+use crate::{clients::FileSystemClient, Properties};
 use azure_core::prelude::*;
 use azure_core::{
     headers::add_optional_header, headers::add_optional_header_ref, AppendToUrlQuery,
@@ -9,37 +9,32 @@ use http::status::StatusCode;
 use std::convert::TryInto;
 
 #[derive(Debug, Clone)]
-pub struct SetFileSystemPropertiesBuilder<'a> {
+pub struct CreateFileSystemBuilder<'a> {
     file_system_client: &'a FileSystemClient,
-    if_modified_since_condition: Option<IfModifiedSinceCondition>,
     client_request_id: Option<ClientRequestId<'a>>,
     timeout: Option<Timeout>,
     properties: Option<&'a Properties<'a, 'a>>,
 }
 
-impl<'a> SetFileSystemPropertiesBuilder<'a> {
-    pub(crate) fn new(
-        file_system_client: &'a FileSystemClient,
-        properties: Option<&'a Properties<'a, 'a>>,
-    ) -> Self {
+impl<'a> CreateFileSystemBuilder<'a> {
+    pub(crate) fn new(file_system_client: &'a FileSystemClient) -> Self {
         Self {
             file_system_client,
-            if_modified_since_condition: None,
             client_request_id: None,
             timeout: None,
-            properties,
+            properties: None,
         }
     }
 
     setters! {
-        if_modified_since_condition: IfModifiedSinceCondition => Some(if_modified_since_condition),
         client_request_id: ClientRequestId<'a> => Some(client_request_id),
         timeout: Timeout => Some(timeout),
+        properties: &'a Properties<'a, 'a> => Some(properties),
     }
 
     pub async fn execute(
         &self,
-    ) -> Result<SetFileSystemPropertiesResponse, Box<dyn std::error::Error + Sync + Send>> {
+    ) -> Result<CreateFileSystemResponse, Box<dyn std::error::Error + Sync + Send>> {
         // we clone this so we can add custom
         // query parameters
         let mut url = self.file_system_client.url().clone();
@@ -51,11 +46,10 @@ impl<'a> SetFileSystemPropertiesBuilder<'a> {
 
         let request = self.file_system_client.prepare_request(
             url.as_str(),
-            &Method::PATCH,
+            &Method::PUT,
             &|mut request| {
-                request = add_optional_header(&self.if_modified_since_condition, request);
-                request = add_optional_header_ref(&self.properties, request);
                 request = add_optional_header(&self.client_request_id, request);
+                request = add_optional_header_ref(&self.properties, request);
                 request
             },
             None,
@@ -66,7 +60,7 @@ impl<'a> SetFileSystemPropertiesBuilder<'a> {
         let response = self
             .file_system_client
             .http_client()
-            .execute_request_check_status(request.0, StatusCode::OK)
+            .execute_request_check_status(request.0, StatusCode::CREATED)
             .await?;
 
         Ok((&response).try_into()?)
